数据可视化-学生信息统计以及班级人数统计代码开发 功能完善：无法删除有员工的部门

+++ tlias-web-management/src/main/java/org/ycc/service/impl/DeptServiceImpl.java
@@ -3,6 +3,7 @@ package org.ycc.service.impl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.ycc.mapper.DeptMapper;
+import org.ycc.mapper.EmpMapper;
 import org.ycc.pojo.Dept;
 import org.ycc.service.DeptService;
 
@@ -15,6 +16,9 @@ public class DeptServiceImpl implements DeptService {
     @Autowired
     private DeptMapper deptMapper;
 
+    @Autowired
+    private EmpMapper empMapper;
+
     @Override
     public List<Dept> findAll() {
         return deptMapper.findAll();
@@ -22,6 +26,10 @@ public class DeptServiceImpl implements DeptService {
 
     @Override
     public void deleteById(Integer id) {
+        Long count = empMapper.countByDeptId(id);
+        if (count > 0) {
+            throw new RuntimeException("对不起，当前部门下有员工，不能直接删除！");
+        }
         deptMapper.deleteById(id);
     }
